python_models/requirements.txt
conda-pack==0.6.0
numpy==1.23.1
Pillow==9.2.0
